Added variance, min and max to and renamed daily BFG harvester (was specific to precip). These changes generalize the daily mean precip harvester so that it can be used for any variable coming from the bfg_ background forecast files
