+++ src/score_hv/hv_registry.py
@@ -1,23 +1,22 @@
-"""
-Copyright 2022 NOAA
+"""Copyright 2023 NOAA
 All rights reserved.
 
 Collection of methods to facilitate file/object retrieval
-
 """
+
 from collections import namedtuple
 
 from score_hv.harvesters.innov_netcdf import InnovStatsCfg, InnovStatsHv
 from score_hv.harvesters.obs_log import ObsInfoCfg, ObsInfoHv
 from score_hv.harvesters.inc_logs import LogIncCfg, LogIncHv
-from score_hv.harvesters.global_bucket_precip_ave import GlobalBucketPrecipRateConfig, GlobalBucketPrecipRateHv
+from score_hv.harvesters.daily_bfg import DailyBFGConfig, DailyBFGHv
 
 NAMED_TUPLES_LIST = 'tuples_list'
 PANDAS_DATAFRAME = 'pandas_dataframe'
 INNOV_NETCDF = 'innov_stats_netcdf'
 OBS_INFO_LOG = 'obs_info_log'
 INC_LOGS = 'inc_logs'
-GLOBAL_BUCKET_PRECIP_AVE = 'global_bucket_precip_ave'
+DAILY_BFG = 'daily_bfg'
 
 Harvester = namedtuple('Harvester', ('name', 'config_handler', 'data_parser'),)
 
@@ -37,9 +36,8 @@ harvester_registry = {INNOV_NETCDF: Harvester(
                           'log files',
                           LogIncCfg,
                           LogIncHv),
-                      GLOBAL_BUCKET_PRECIP_AVE: Harvester(
-                          'Global mean bucket surface precipitation rate. ' 
-                          'Adding from background forecast data',
-                          GlobalBucketPrecipRateConfig,
-                          GlobalBucketPrecipRateHv)
+                      DAILY_BFG: Harvester(
+                          'Daily mean values from background forecast data',
+                          DailyBFGConfig,
+                          DailyBFGHv)
                       }
